refactor: move chatbot url to chatbot config
BREAKING CHANGE: new api to chatbot
