+++ src/lib/bot/chatbot.js
@@ -1,6 +1,7 @@
-import chatbotLogo from "../assets/images/bacoor-mayor-vmayor.jpg";
+import chatbotLogo from "../../assets/images/bacoor-mayor-vmayor.jpg";
 
 export const chatbot = {
+  url: "https://chatbot-api-0zup.onrender.com/api/chatbot/query",
   logo: chatbotLogo,
   name: "Viviane",
   nickName: "Viviane",
